Add configure and mkconfig.tmp to work around a
name conflict that need not exist.  It was a conflict between the
osh and .oshrc files, found in the Osh Project [1], and files with
the same names, found in the OMake build system [2].  Fixed the
problem by moving osh and its .oshrc to etsh (enhanced Thompson
shell) and .etshrc I did.  The benefit is that omake and osh can
now be installed side by side on the same machine without conflict.

1. https://v6shell.org/
2.  http://omake.metaprl.org/